refactor: rearrange iface package
