+++ iface/iface_darwin.go
@@ -7,9 +7,10 @@ import (
 	"strings"
 )
 
-//const (
-//	interfacePrefix = "utun"
-//)
+// Create Creates a new Wireguard interface, sets a given IP and brings it up.
+func Create(iface string, address string) error {
+	return CreateWithUserspace(iface, address)
+}
 
 // assignAddr Adds IP address to the tunnel interface and network route based on the range provided
 func assignAddr(address string, ifaceName string) error {
