+++ iface/iface_windows.go
@@ -8,10 +8,15 @@ import (
 	"net"
 )
 
+// Create Creates a new Wireguard interface, sets a given IP and brings it up.
+func Create(iface string, address string) error {
+	return CreateWithUserspace(iface, address)
+}
+
 // assignAddr Adds IP address to the tunnel interface and network route based on the range provided
 func assignAddr(address string, ifaceName string) error {
 
-	nativeTunDevice := tunDevice.(*tun.NativeTun)
+	nativeTunDevice := tunIface.(*tun.NativeTun)
 	luid := winipcfg.LUID(nativeTunDevice.LUID())
 
 	ip, ipnet, _ := net.ParseCIDR(address)
